backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.6.0
pydantic[email]==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pymongo==4.9.1
python-multipart==0.0.9
apscheduler==3.10.4
websockets==12.0

# Phase 3 — AI dependencies
pyyaml==6.0.2
httpx==0.27.2
